Submission ready
* api added
* cleaned up some files

+++ serve/predict.py
@@ -90,6 +90,6 @@ def predict_fn(input_data, model):
     with torch.no_grad():
         output = model.forward(data)
         
-    result = np.round(output.numpy()).astype(int)
+    result = np.round(output.numpy())
 
     return result
